+++ ladim/__init__.py
@@ -1,3 +1,3 @@
-__version__ = '1.3.5'
+__version__ = '2.0.0'
 
 from .main import main, run